// backend/src/common/enums/department-permission.enum.ts
export enum DepartmentPermissionEnum {
  VIEW_ONLY = 1,
  UPDATE_PERMISSION_IN_DEPARTMENT = 2,
  VIEW_ALL_EMPLOYEE_IN_SPECIFIC_DEPARTMENT = 3,
  INVITE_EMPLOYEE_TO_SPECIFIC_DEPARTMENT = 4,
  UPDATE_EMPLOYEE_IN_SPECIFIC_DEPARTMENT = 5,
  DELETE_EMPLOYEE_IN_SPECIFIC_DEPARTMENT = 6,
  UPDATE_DEPARTMENT = 7,
  DELETE_DEPARTMENT = 8,
}